hmm float column type better?

+++ db/migrate/20111218005040_track_rate.rb
@@ -1,6 +1,6 @@
 class TrackRate < ActiveRecord::Migration
   def up
-    add_column :swims, :rate, :double
+    add_column :swims, :rate, :float
 
     Swim.measured.each do |s|
       s.calculate_rate
